Skip <Arduino.h> while parsing
I'm working on personal project and use VSCode. I have to include #include <Arduino.h> to get the auto complete to work right. This unfortunately means I will get the following in my stderr when using the script. 
```
  NOT FOUND: Arduino.h
```   
It still works, but I do not consider this an error. Anyway my PR here just hardcodes to skip it since it will never be a requirement.